feat(feedback): refactor project structure and update dependencies for ml integration
