+++ dg_projects/ml/build.yaml
@@ -1,0 +1,6 @@
+---
+# Build context is set to repository root to access packages/ol-orchestrate-lib
+directory: ../..
+# The Dockerfile is located in the ml directory
+dockerfile: dg_projects/ml/Dockerfile
+registry: mitodl/dagster-ml
