+++ dg_projects/ml/ml/definitions.py
@@ -1,5 +1,3 @@
-import os
-
 from dagster import (
     Definitions,
     define_asset_job,
@@ -7,8 +5,9 @@ from dagster import (
 from dagster_aws.s3 import S3Resource
 from dagster_iceberg.config import IcebergCatalogConfig
 from dagster_iceberg.io_manager.polars import PolarsIcebergIOManager
-from feedback_clustering.assets.feedback_clustering import feedback_redacted
-from feedback_clustering.resources.llm import LLMClientFactory
+from ml.assets.feedback_redacted import feedback_redacted
+from ml.assets.risk_probability import student_risk_probability
+from ml.resources.llm import LLMClientFactory
 from ol_orchestrate.lib.constants import DAGSTER_ENV, VAULT_ADDRESS
 from ol_orchestrate.lib.dagster_helpers import (
     default_file_object_io_manager,
@@ -21,7 +20,7 @@ from ol_orchestrate.lib.utils import (
     unauthenticated_vault,
 )
 
-init_sentry("feedback_clustering")
+init_sentry("ml")
 
 # Initialize vault with resilient loading
 try:
@@ -37,12 +36,10 @@ except Exception as e:  # noqa: BLE001 (resilient loading)
     vault = unauthenticated_vault(VAULT_ADDRESS)
     vault_authenticated = False
 
-if DAGSTER_ENV == "dev":
-    database_name = (
-        f"ol_warehouse_production_{os.environ.get('DBT_SCHEMA_SUFFIX')}_intermediate"
-    )
-else:
-    database_name = "ol_warehouse_production_intermediate"
+data_export_job = define_asset_job(
+    name="student_risk_probability_data_export_job",
+    selection=[student_risk_probability],
+)
 
 feedback_redacted_job = define_asset_job(
     name="feedback_redacted_job",
@@ -52,6 +49,15 @@ feedback_redacted_job = define_asset_job(
 # Create unified definitions
 defs = Definitions(
     resources={
+        # namespace intentionally omitted here: dagster's DbIOManager schema
+        # precedence is output metadata > io_manager namespace > asset key
+        # prefix > "public", and the AssetKey prefix (["reporting", ...] /
+        # ["intermediate", ...]) is a literal Glue database name, not env-aware
+        # -- it does not expand to ol_warehouse_production_<suffix>_reporting.
+        # Each asset sets its own env-aware `metadata={"schema": ...}` instead
+        # (see risk_probability.py / feedback_redacted.py), which is what
+        # actually determines its write target; a shared namespace here would
+        # force both assets to the same schema.
         "io_manager": PolarsIcebergIOManager(
             name="iceberg_io_manager",
             config=IcebergCatalogConfig(
@@ -68,7 +74,6 @@ defs = Definitions(
                     "s3.request-timeout": "120",
                 }
             ),
-            namespace=database_name,
             reader_override="pyiceberg",
         ),
         "s3file_io_manager": default_file_object_io_manager(
@@ -80,6 +85,6 @@ defs = Definitions(
         "s3": S3Resource(),
         "llm": LLMClientFactory(vault=vault),
     },
-    assets=with_failure_hooks([feedback_redacted]),
-    jobs=[feedback_redacted_job],
+    assets=with_failure_hooks([student_risk_probability, feedback_redacted]),
+    jobs=[data_export_job, feedback_redacted_job],
 )
