+++ dg_projects/ml/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "feedback_clustering"
+name = "ml"
 requires-python = ">=3.14,<3.15"
 version = "0.1.0"
 dependencies = [
@@ -10,7 +10,7 @@ dependencies = [
     "dagster-postgres>=0.27.13",
     "dagster-iceberg>=0.3.8",
     "httpx2 >= 2.2.0",
-    "scikit-learn>=1.3",
+    "scikit-learn>=1.2.2",
     "ol-orchestrate-lib",
     "polars<1.44",
     "fenic",
@@ -19,10 +19,11 @@ dependencies = [
     "presidio-analyzer",
     "presidio-anonymizer",
     "spacy>=3.7",
+    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl",
 ]
 
 [project.entry-points]
-"dagster.components" = { feedback_clustering = "feedback_clustering.lib"}
+"dagster.components" = { ml = "ml.lib"}
 
 [dependency-groups]
 dev = [
@@ -32,24 +33,24 @@ dev = [
 ]
 
 [tool.pytest.ini_options]
-testpaths = ["feedback_clustering_tests"]
+testpaths = ["ml_tests"]
 
 [build-system]
 requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.dagster]
-module_name = "feedback_clustering.definitions"
-project_name = "feedback_clustering"
+module_name = "ml.definitions"
+project_name = "ml"
 
 [tool.dg]
 directory_type = "project"
 
 [tool.dg.project]
-root_module = "feedback_clustering"
+root_module = "ml"
 
 [tool.setuptools.packages.find]
-exclude=["feedback_clustering_tests"]
+exclude=["ml_tests"]
 
 [tool.uv.sources]
 ol-orchestrate-lib = { path = "../../packages/ol-orchestrate-lib", editable = true }
