--- dg_projects/student_risk_probability/Dockerfile
@@ -1,46 +1,0 @@
-# Based on: https://github.com/astral-sh/uv-docker-example/blob/main/multistage.Dockerfile
-
-# First, build the application in the `/app` directory.
-FROM ghcr.io/astral-sh/uv:python3.14-bookworm-slim AS builder
-ENV UV_COMPILE_BYTECODE=1 UV_LINK_MODE=copy
-
-# Disable Python downloads, because we want to use the system interpreter
-# across both images. If using a managed Python version, it needs to be
-# copied from the build image into the final image; see `standalone.Dockerfile`
-# for an example.
-ENV UV_PYTHON_DOWNLOADS=0
-
-# Install build dependencies required for compiling C extensions (e.g., pyroaring)
-RUN apt-get update && apt-get install -y --no-install-recommends \
-    build-essential \
-    && rm -rf /var/lib/apt/lists/*
-
-WORKDIR /app
-
-# Copy the entire workspace structure needed for uv
-COPY packages/ol-orchestrate-lib packages/ol-orchestrate-lib
-COPY dg_projects/student_risk_probability dg_projects/student_risk_probability
-
-# Install dependencies from the student_risk_probability project
-WORKDIR /app/dg_projects/student_risk_probability
-RUN --mount=type=cache,target=/root/.cache/uv \
-    uv venv --relocatable && \
-    uv sync --frozen --no-install-project --no-dev --no-editable
-
-# Sync the project itself
-RUN --mount=type=cache,target=/root/.cache/uv \
-    uv sync --frozen --no-dev --no-editable
-
-
-# Then, use a final image without uv
-FROM python:3.14-slim-bookworm
-# It is important to use the image that matches the builder, as the path to the
-# Python executable must be the same.
-
-# Copy the application from the builder
-COPY --from=builder /app/dg_projects/student_risk_probability /app
-
-# Place executables in the environment at the front of the path
-ENV PATH="/app/.venv/bin:$PATH"
-
-WORKDIR /app
