--- dg_projects/student_risk_probability/build.yaml
@@ -1,6 +1,0 @@
----
-# Build context is set to repository root to access packages/ol-orchestrate-lib
-directory: ../..
-# The Dockerfile is located in the student_risk_probability directory
-dockerfile: dg_projects/student_risk_probability/Dockerfile
-registry: mitodl/dagster-student_risk_probability
