--- dg_projects/student_risk_probability/pyproject.toml
@@ -1,58 +1,0 @@
-[project]
-name = "student_risk_probability"
-requires-python = ">=3.14,<3.15"
-version = "0.1.0"
-dependencies = [
-    "dagster>=1.12.10",
-    "dagster-dbt ~=0.29.0",
-    "dagster-aws ~=0.29.0",
-    "dagster-k8s>=0.27.15",
-    "dagster-postgres>=0.27.13",
-    "dagster-iceberg>=0.3.8",
-    "httpx2 >= 2.2.0",
-    "scikit-learn>=1.2.2",
-    "ol-orchestrate-lib",
-    "polars<1.44",
-]
-
-[project.entry-points]
-"dagster.components" = { student_risk_probability = "student_risk_probability.lib"}
-
-[dependency-groups]
-dev = [
-    "dagster-dg-cli>=1.12.10",
-    "dagster-webserver",
-]
-
-[build-system]
-requires = ["setuptools"]
-build-backend = "setuptools.build_meta"
-
-[tool.dagster]
-module_name = "student_risk_probability.definitions"
-project_name = "student_risk_probability"
-
-[tool.dg]
-directory_type = "project"
-
-[tool.dg.project]
-root_module = "student_risk_probability"
-
-[tool.setuptools.packages.find]
-exclude=["student_risk_probability_tests"]
-
-[tool.uv.sources]
-ol-orchestrate-lib = { path = "../../packages/ol-orchestrate-lib", editable = true }
-
-# Override dagster-iceberg's <0.11 upper bound until upstream loosens it.
-# See https://github.com/dagster-io/community-integrations/pull/289
-#
-# Pin grpcio-health-checking below 1.82.0: that release bundles health_pb2.py
-# pre-generated with protobuf 7.35.0 gencode while declaring protobuf>=6.33.5,
-# so it crashes at import time under dagster's required protobuf<7 (upstream
-# metadata/gencode mismatch in the 1.82.0 release).
-[tool.uv]
-override-dependencies = [
-    "pyiceberg>=0.11.1",
-    "grpcio-health-checking<1.82.0",
-]
